+++ lib/Icon/raw/IntoInventory.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+export const IntoInventory = () => (
+    <svg
+       xmlns="http://www.w3.org/2000/svg"
+       width="18"
+       height="18"
+       viewBox="0 0 4.7624999 4.7625001"
+       version="1.1"
+       >
+    <path
+       className="cogi-icon-raw"
+       d="m 0.26458333,0.26458333 2.91041677,2e-8 L 3.7041667,0.79375 V 3.175 L 2.9104167,2.38125 H 3.1750001 V 1.0583333 l -0.2645834,-0.26458328 -2.64583337,-3e-8 z"
+       />
+    <path
+       className="cogi-icon-raw"
+       d="m 4.4979166,2.9104166 v 1.5875 H 0.26458333 v -1.5875 h 0.52916666 v 0.79375 L 1.0583333,3.96875 H 3.7041666 L 3.96875,3.7041666 v -0.79375 z"
+       />
+    </svg>
+);