+++ lib/Icon/raw/Plus.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+export const Plus = () => (
+    <svg
+       xmlns="http://www.w3.org/2000/svg"
+       width="18"
+       height="18"
+       viewBox="0 0 4.7624999 4.7625001"
+       version="1.1"
+       >
+        <path
+            className="cogi-icon-raw"
+            d="M 2.1166666,0.26458333 V 2.1166666 H 0.26458333 V 2.6458333 H 2.1166666 V 4.4979166 H 2.6458333 V 2.6458333 H 4.4979166 V 2.1166666 H 2.6458333 V 0.26458333 Z"
+            />
+    </svg>
+);