+++ lib/PaintCard/PaintCard.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { PaintCard } from "../index";
+import { ContextualControls, PaintCard } from "../index";
 import { Paint } from '@cogisphere/cogi-colors';
 
 export default {
@@ -16,4 +16,22 @@ const paint = new Paint({ hex: '#dd55cd', paintCode: "test-something", name: "Te
 export const Default = Template.bind({ });
 Default.args = {
     paint
+};
+
+export const WithControls = Template.bind({ });
+WithControls.args = {
+    paint,
+    controls: (<>
+      <button>Click</button>  
+      <span>Controls</span>
+    </>)
+};
+
+export const WithContextualControls = Template.bind({ });
+WithContextualControls.args = {
+    paint,
+    controls: (<ContextualControls items={[
+        { icon: "plus" },
+        { icon: "cross" }
+    ]}/>)
 };